requirements.txt
flask-sqlalchemy
psycopg2
python-dotenv
gunicorn
Flask-WTF
email-validator
faker
